requirements.txt
absl-py==0.12.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
Cython @ file:///tmp/build/80754af9/cython_1614014840177/work
decorator==4.4.2
einops==0.3.0
filelock==3.0.12
gdown==3.12.2
google-auth==1.28.0
google-auth-oauthlib==0.4.4
grpcio==1.36.1
idna==2.10
imageio==2.9.0
importlib-metadata==3.10.0
jpeg4py==0.1.4
jsonpatch==1.32
jsonpointer==2.1
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282414123/work
Markdown==3.3.4
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1613407855456/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
networkx==2.5
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.1.48
pandas==1.2.3
Pillow @ file:///tmp/build/80754af9/pillow_1615057347459/work
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks==1.7.1
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scipy @ file:///tmp/build/80754af9/scipy_1616700973326/work
Shapely==1.6.4.post2
six @ file:///tmp/build/80754af9/six_1605205313296/work
tb-nightly==2.5.0a20210329
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tifffile==2021.3.17
tikzplotlib==0.9.8
torch==1.5.1
torchfile==0.1.0
torchvision==0.6.0a0+35d732a
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
tqdm @ file:///tmp/build/80754af9/tqdm_1615925068909/work
typing-extensions==3.7.4.3
urllib3==1.26.4
visdom==0.1.8.9
websocket-client==0.58.0
Werkzeug==1.0.1
wget==3.2
yacs==0.1.8
zipp==3.4.1
